requirements.txt
python-dotenv
requests
schedule
xmltodict